requirements/base.txt
# Shared dependencies for both local and production environments
argon2-cffi-bindings==21.2.0
argon2-cffi==25.1.0
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
django-allauth==65.9.0
django-ckeditor==6.7.3
django-crispy-forms==2.4
django-environ==0.12.0
django-js-asset==3.1.2
django-model-utils==5.0.0
django-postcss-modules==0.1.2
django-redis==6.0.0
django-tailwind==4.0.1
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
django==5.2.3
hiredis==3.2.1
idna==3.10
markdown2==2.5.3
markdown==3.8.2
packaging==25.0
pillow==11.2.1
py-mini-racer==0.6.0
pycparser==2.22
python-slugify==8.0.4
pytz==2025.2
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
